Successfully inject contentResolver
