+++ InTouch/app/src/main/java/com/n8/intouch/addeventscreen/AddEventInteractorImpl.kt
@@ -1,0 +1,10 @@
+package com.n8.intouch.addeventscreen
+
+import android.content.Context
+
+/**
+ * Created by n8 on 11/30/15.
+ */
+class AddEventInteractorImpl(var context: Context) : AddEventInteractor {
+
+}
