+++ InTouch/app/src/main/java/com/n8/intouch/addeventscreen/AddEventPresenter.kt
@@ -1,0 +1,7 @@
+package com.n8.intouch.addeventscreen
+
+/**
+ * Created by n8 on 11/30/15.
+ */
+interface AddEventPresenter {
+}
