+++ InTouch/app/src/main/java/com/n8/intouch/addeventscreen/AddEventPresenterImpl.kt
@@ -1,0 +1,8 @@
+package com.n8.intouch.addeventscreen
+
+/**
+ * Created by n8 on 11/30/15.
+ */
+class AddEventPresenterImpl(var view: AddEventView, var interactor: AddEventInteractor) : AddEventPresenter{
+
+}
